+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/module/farming/livestock/AnimalMixin_Livestock.java
@@ -34,19 +34,19 @@ public abstract class AnimalMixin_Livestock extends AgeableMob {
     @Definition(id = "inLove", field = "Lnet/minecraft/world/entity/animal/Animal;inLove:I")
     @Expression("this.inLove > 0")
     @ModifyExpressionValue(method = "aiStep", at = @At(value = "MIXINEXTRAS:EXPRESSION"))
-    private boolean iguanatweaksreborn$checkIfServerSideBeforeSendingHearts(boolean original) {
+    private boolean insanesurvivaloverhaulcheckIfServerSideBeforeSendingHearts(boolean original) {
         return original && !this.level().isClientSide;
     }
 
     @WrapOperation(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/Level;addParticle(Lnet/minecraft/core/particles/ParticleOptions;DDDDDD)V"))
-    private void iguanatweaksreborn$sendLoveHeartsToClients(Level instance, ParticleOptions pParticleData, double pX, double pY, double pZ, double pXSpeed, double pYSpeed, double pZSpeed, Operation<Void> original) {
+    private void insanesurvivaloverhaulsendLoveHeartsToClients(Level instance, ParticleOptions pParticleData, double pX, double pY, double pZ, double pXSpeed, double pYSpeed, double pZSpeed, Operation<Void> original) {
         ((ServerLevel) instance).sendParticles(pParticleData, Livestock.getRandomXWithin(this, 0.2f), this.getRandomY() + 0.5, Livestock.getRandomZWithin(this,0.2f), 1, 0, 0, 0, 0.1f);
     }
 
     @Definition(id = "pPlayer", local = @Local(type = Player.class, argsOnly = true))
     @Expression("pPlayer != null")
     @WrapOperation(method = "setInLove", at = @At(value = "MIXINEXTRAS:EXPRESSION"))
-    private boolean iguanatweaksreborn$allowSetLoveCauseAnyway(Object left, Object right, Operation<Boolean> original) {
+    private boolean insanesurvivaloverhaulallowSetLoveCauseAnyway(Object left, Object right, Operation<Boolean> original) {
         boolean originalR = original.call(left, right);
         if (!originalR)
             this.loveCause = null;