+++ changelog.md
@@ -1,5 +1,5 @@
 ## Upcoming
-* Added back Farming features: Crops, Bone Meal
+* Added back Farming features: Crops, Bone Meal, Livestock, Hoes
   * Rich farmland has been disabled by default
   * Crops can now grow with bone meal when farmland is not moist
 * Fixed Water Fall Damage feature not working properly